Delete `LocalHarness`.
